build: use extra vite config for lib
